Moved linter functions from main module
