+++ lib/linter-docblock-python.js
@@ -5,6 +5,26 @@ import {get_all_docblocks} from './docblock-python.js';
 
 this.get_all_docblocks = get_all_docblocks;
 
+export function provideLinter() {
+  return {
+    name: 'docblock-python',
+    scope: 'file', // or 'project'
+    lintsOnChange: true, // or true
+    grammarScopes: ['source.python'],
+    lint(textEditor) {
+      const editorPath = textEditor.getPath();
+      return new Promise(function(resolve) {
+        resolve((lint_docblocks() || [])
+          .map((missing) => {
+            return formatLint(missing.pos,
+              missing.par, missing.type, editorPath);
+          })
+        );
+      });
+    },
+  };
+}
+
 function formatLint(msg_pos, par_name, par_type, editorPath) {
   let type_text;
   if (par_type === 'attr') {
