+++ lib/utils.js
@@ -16,9 +16,7 @@ warn.nothing_to_do = () => {
   );
 };
 
-export function generate_docblock() {
-  let editor = atom.workspace.getActiveTextEditor();
-  let docblock;
+function get_atom_config() {
   // Load settings:
   let [tab_length, tab_type] = ['editor.tabLength', 'editor.softTabs']
     .map((q) => atom.config.get(q, {scope: ['source.python']}));
@@ -75,6 +73,14 @@ export function generate_docblock() {
     template: template,
     tabs: '',
   };
+  return options;
+}
+
+export function generate_docblock() {
+  let editor = atom.workspace.getActiveTextEditor();
+  let docblock;
+
+  let options = get_atom_config();
 
   if (editor) {
     let language = editor.getGrammar().name;
@@ -98,7 +104,9 @@ export function generate_docblock() {
         editor.moveToFirstCharacterOfLine();
         pos = editor.getCursorBufferPosition();
         let col = pos.column;
-        let tabs = tab_type.repeat((col / tab_length) + indent);
+        let tabs = options.tab_type.repeat(
+          (col / options.tab_length) + options.indent
+        );
         options.tabs = tabs;
         editor.selectToEndOfLine();
         let query = editor.getSelectedText();
@@ -159,7 +167,7 @@ export function generate_docblock() {
   }
 }
 
-export function get_header(options, start, lines, accum='') {
+function get_header(options, start, lines, accum='') {
   let editor = atom.workspace.getActiveTextEditor();
   let query;
   if (editor) {
@@ -635,3 +643,144 @@ function getFunctionReturnType(query) {
 
   return return_type;
 }
+
+function get_def(start, direction) {
+  let query;
+  let def;
+  if (editor = atom.workspace.getActiveTextEditor()) {
+    if (start.row < 0) {
+      return {};
+    }
+    query = editor.lineTextForBufferRow(start.row);
+    let should_stop = ['return', 'yield']
+      .map((x) => query.search(x))
+      .some((x) => x >= 0);
+    if (should_stop) {
+      return {};
+    }
+    let logic = main.accepted_starts
+      .map((x) => query.search(x + ' '))
+      .some((x) => x >= 0);
+    if (logic) {
+      let header = get_header(start, 0);
+      def = {def: header, pos: start};
+    } else {
+      if (direction === 'up') {
+        def = get_def({row: start.row - 1, column: start.column}, 'up');
+      } else if (direction === 'down') {
+        def = get_def({row: start.row + 1, column: start.column}, 'down');
+      }
+    };
+    if (def === undefined) {
+      return {};
+    };
+    if (Object.keys(def).length) {
+      return {def: def.def, pos: def.pos};
+    } else {
+      return {};
+    };
+  };
+}
+
+export function get_all_docblocks() {
+  let options = get_atom_config();
+  let editor = atom.workspace.getActiveTextEditor();
+  if (editor) {
+    let allText = editor.getText();
+    let dbRegex = new RegExp('"""([^])*?"""', 'g');
+    let match = dbRegex.exec(allText);
+    let allBlocks = [];
+    let last_end = -1;
+    while (match) {
+      let start = editor.buffer.positionForCharacterIndex(match.index);
+      if (atom.config.get('docblock-python.indent')) {
+        start.column = start.column - options.tab_length;
+      };
+      let def = get_def({row: start.row, column: start.column}, 'up');
+      if (Object.keys(def).length && last_end < def.pos.row) {
+        let block = {
+          docblock: match[0],
+          pos_start_char: match.index,
+          pos_start: start,
+          pos_end_char: dbRegex.lastIndex,
+          pos_end: editor.buffer.positionForCharacterIndex(dbRegex.lastIndex
+            + 1),
+          def: def.def[0],
+          def_lines: def.def[1],
+          def_pos: def.pos,
+        };
+        last_end = block.pos_end.row;
+        allBlocks.push(block);
+      };
+      match = dbRegex.exec(allText);
+    }
+    return allBlocks;
+  };
+}
+
+// function get_docblock() {
+//   let editor = atom.workspace.getActiveTextEditor();
+//   if (editor) {
+//     let start_pos = editor.getCursorBufferPosition();
+//     editor.moveToEndOfLine();
+//     let pos_for_up = editor.getCursorBufferPosition();
+//     let text_up = scan_up(pos_for_up, 0);
+//     editor.setCursorBufferPosition(start_pos);
+//     editor.moveToBeginningOfLine();
+//     let pos_for_down = editor.getCursorBufferPosition();
+//     let text_down = scan_down(pos_for_down, 0);
+//     if ((text_up !== null) && (text_down !== null)) {
+//       lines_up = text_up[0].split('\n');
+//       lines_down = text_down[0].split('\n').slice(1);
+//       return lines_up.join('\n') + lines_down.join('\n');
+//     }
+//   };
+// }
+//
+// function scan_up(start, lines) {
+//   let editor = atom.workspace.getActiveTextEditor();
+//   let query;
+//   if (editor) {
+//     editor.setCursorBufferPosition(start);
+//     editor.moveUp(lines);
+//     editor.moveToBeginningOfLine();
+//     editor.selectToBufferPosition(start);
+//     let query = editor.getSelectedText();
+//     if (!(/"""/.test(query))) {
+//       let logic = main.stop_words
+//         .map((x) => query.search(x))
+//         .some((x) => x >= 0);
+//       if (logic) {
+//         return null;
+//       } else {
+//         tmp = scan_up(start, lines + 1);
+//         query = tmp[0];
+//         lines = tmp[1];
+//       };
+//     };
+//   };
+//   return [query, lines];
+// }
+//
+// function scan_down(start, lines) {
+//   let editor = atom.workspace.getActiveTextEditor();
+//   let query;
+//   if (editor) {
+//     editor.setCursorBufferPosition(start);
+//     editor.moveDown(lines);
+//     editor.moveToEndOfLine();
+//     editor.selectToBufferPosition(start);
+//     let query = editor.getSelectedText().trim();
+//     if (!(/"""/.test(query))) {
+//       let logic = main.stop_words
+//         .map((x) => query.search(x))
+//         .some((x) => x >= 0);
+//
+//       if (logic) return null;
+//       tmp = scan_down(start, lines + 1);
+//       query = tmp[0];
+//       lines = tmp[1];
+//     };
+//   };
+//   return [query, lines];
+// }
